Address self-review: extract shared buildCustomFieldProps, add maxCount, fix checkbox required and MIME validation
