+++ clients/privacy-center/components/modals/consent-request-modal/ConsentRequestForm.tsx
@@ -1,14 +1,11 @@
-import { Button, Flex, Form, Input, Text, UploadFile } from "fidesui";
+import { Button, Flex, Form, Input, Text } from "fidesui";
 import React, { useEffect } from "react";
 
-import CustomFieldRenderer, {
-  CustomFieldRendererProps,
-} from "~/components/common/CustomFieldRenderer";
+import { buildCustomFieldProps } from "~/components/common/buildCustomFieldProps";
+import CustomFieldRenderer from "~/components/common/CustomFieldRenderer";
 import { ModalViews } from "~/components/modals/types";
 import { PhoneInput } from "~/components/phone-input";
 import { useConfig } from "~/features/common/config.slice";
-import { CustomConfigField } from "~/types/config";
-import { FormFieldValue } from "~/types/forms";
 
 import useConsentRequestForm from "./useConsentRequestForm";
 
@@ -112,98 +109,31 @@ const ConsentRequestForm = ({
             />
           </Form.Item>
         )}
-        {Object.entries(customPrivacyRequestFields).map(([key, item]) => {
-          const customFieldProps = (
-            value: FormFieldValue,
-            fieldConfig: CustomConfigField,
-          ): CustomFieldRendererProps => {
-            const sharedProps = {
-              fieldKey: key,
-              onBlur: () => handleBlur({ target: { name: key } }),
-              error:
-                touched[key] && errors[key]
-                  ? (errors[key] as string)
-                  : undefined,
-            };
-
-            switch (fieldConfig.field_type) {
-              case "multiselect":
-              case "checkbox_group": {
-                let arrayValue: string[];
-                if (typeof value === "string") {
-                  arrayValue = [value];
-                } else if (Array.isArray(value)) {
-                  arrayValue = value as string[];
-                } else {
-                  arrayValue = [];
-                }
-                return {
-                  ...fieldConfig,
-                  ...sharedProps,
-                  value: arrayValue,
-                  onChange: (v: Array<string>) => {
-                    setFieldValue(key, v);
-                  },
-                };
-              }
-              case "checkbox":
-                return {
-                  ...fieldConfig,
-                  ...sharedProps,
-                  value: Boolean(value),
-                  onChange: (v: boolean) => {
-                    setFieldValue(key, v);
-                  },
-                };
-              case "file":
-                return {
-                  ...fieldConfig,
-                  ...sharedProps,
-                  value: Array.isArray(value) ? (value as UploadFile[]) : [],
-                  onChange: (fileList: UploadFile[]) => {
-                    setFieldValue(key, fileList);
-                  },
-                };
-              default: {
-                let stringValue: string;
-                if (typeof value === "string") {
-                  stringValue = value;
-                } else if (Array.isArray(value) && value.length > 0) {
-                  stringValue = value[0] as string;
-                } else {
-                  stringValue = "";
-                }
-                return {
-                  ...fieldConfig,
-                  ...sharedProps,
-                  value: stringValue,
-                  onChange: (v: string) => {
-                    setFieldValue(key, v);
-                  },
-                };
-              }
+        {Object.entries(customPrivacyRequestFields).map(([key, item]) => (
+          <Form.Item
+            key={key}
+            id={key}
+            validateStatus={
+              touched[key] && Boolean(errors[key]) ? "error" : undefined
             }
-          };
-
-          return (
-            <Form.Item
-              key={key}
-              id={key}
-              validateStatus={
-                touched[key] && Boolean(errors[key]) ? "error" : undefined
-              }
-              help={touched[key] && (errors[key] as string)}
-              required={item.required !== false}
-              hasFeedback={
-                item.field_type === "text" && touched[key] && !!errors[key]
-              }
-              label={item.label}
-              htmlFor={key}
-            >
-              <CustomFieldRenderer {...customFieldProps(values[key], item)} />
-            </Form.Item>
-          );
-        })}
+            help={touched[key] && (errors[key] as string)}
+            required={item.required !== false}
+            hasFeedback={
+              item.field_type === "text" && touched[key] && !!errors[key]
+            }
+            label={item.label}
+            htmlFor={key}
+          >
+            <CustomFieldRenderer
+              {...buildCustomFieldProps(key, values[key], item, {
+                setFieldValue,
+                handleBlur,
+                touched,
+                errors,
+              })}
+            />
+          </Form.Item>
+        ))}
         <Flex justify="stretch" gap="medium">
           <Button type="default" variant="outlined" onClick={onClose} block>
             {config.consent?.button.cancelButtonText || "Cancel"}
