+++ clients/privacy-center/components/modals/privacy-request-modal/PrivacyRequestForm.tsx
@@ -1,14 +1,12 @@
-import { Button, Flex, Form, Input, Text, UploadFile } from "fidesui";
+import { Button, Flex, Form, Input, Text } from "fidesui";
 import React from "react";
 
 import { isFieldVisible } from "~/common/visibility";
-import CustomFieldRenderer, {
-  CustomFieldRendererProps,
-} from "~/components/common/CustomFieldRenderer";
+import { buildCustomFieldProps } from "~/components/common/buildCustomFieldProps";
+import CustomFieldRenderer from "~/components/common/CustomFieldRenderer";
 import { ModalViews } from "~/components/modals/types";
 import { PhoneInput } from "~/components/phone-input";
-import { CustomConfigField, PrivacyRequestOption } from "~/types/config";
-import { FormFieldValue } from "~/types/forms";
+import { PrivacyRequestOption } from "~/types/config";
 
 import usePrivacyRequestForm, { OrderedField } from "./usePrivacyRequestForm";
 
@@ -54,66 +52,7 @@ const PrivacyRequestForm = ({
     return null;
   }
 
-  const buildCustomFieldProps = (
-    key: string,
-    value: FormFieldValue,
-    fieldConfig: CustomConfigField,
-  ): CustomFieldRendererProps => {
-    const sharedProps = {
-      fieldKey: key,
-      onBlur: () => handleBlur({ target: { name: key } }),
-      error: touched[key] && errors[key] ? (errors[key] as string) : undefined,
-    };
-    switch (fieldConfig.field_type) {
-      case "multiselect":
-      case "checkbox_group": {
-        let arrayValue: string[];
-        if (typeof value === "string") {
-          arrayValue = [value];
-        } else if (Array.isArray(value)) {
-          arrayValue = value as string[];
-        } else {
-          arrayValue = [];
-        }
-        return {
-          ...fieldConfig,
-          ...sharedProps,
-          value: arrayValue,
-          onChange: (v: Array<string>) => setFieldValue(key, v),
-        };
-      }
-      case "checkbox":
-        return {
-          ...fieldConfig,
-          ...sharedProps,
-          value: Boolean(value),
-          onChange: (v: boolean) => setFieldValue(key, v),
-        };
-      case "file":
-        return {
-          ...fieldConfig,
-          ...sharedProps,
-          value: Array.isArray(value) ? (value as UploadFile[]) : [],
-          onChange: (fileList: UploadFile[]) => setFieldValue(key, fileList),
-        };
-      default: {
-        let stringValue: string;
-        if (typeof value === "string") {
-          stringValue = value;
-        } else if (Array.isArray(value) && value.length > 0) {
-          stringValue = value[0] as string;
-        } else {
-          stringValue = "";
-        }
-        return {
-          ...fieldConfig,
-          ...sharedProps,
-          value: stringValue,
-          onChange: (v: string) => setFieldValue(key, v),
-        };
-      }
-    }
-  };
+  const formContext = { setFieldValue, handleBlur, touched, errors };
 
   const renderField = (field: OrderedField): React.ReactElement | null => {
     if (field.kind === "name") {
@@ -210,7 +149,7 @@ const PrivacyRequestForm = ({
         htmlFor={key}
       >
         <CustomFieldRenderer
-          {...buildCustomFieldProps(key, values[key], item)}
+          {...buildCustomFieldProps(key, values[key], item, formContext)}
         />
       </Form.Item>
     );
